08/10/2024

+++ rf_master/rf_master/bsp_source/inc/timer_driver.h
@@ -16,7 +16,7 @@ typedef struct
 {
       uint16_t tick ;
       uint16_t tick_timer1 ;
-      uint16_t tick_timer3 ;
+      uint32_t tick_timer3 ;
       uint16_t tick_timer2 ;
 }Tick_Timer ;  
 
@@ -42,7 +42,7 @@ void Timer2_ResetTickMs(void);
 
 void SB_Timer3_Init(void);
 void Timer3_SetTickMs(void);
-uint16_t Timer3_GetTickMs(void);
+uint32_t Timer3_GetTickMs(void);
 void Timer3_ResetTickMs(void);
 
 void SB_Timer_Driver_Init(void);
